refactor: rename 'app' folder to 'backend' and add modular vector database structure
