+++ backend/.env.example
@@ -1,0 +1,4 @@
+OPENROUTER_API_KEY=
+STABILITY_API_KEY=
+PINECONE_API_KEY=
+PINECONE_ENV=
