+++ backend/image_generator.py
